Refactoring Marking Schema
