+++ src/routes/(admin)/tests/test-[type]/[id]/schema.ts
@@ -1,22 +1,5 @@
 import { z } from 'zod';
-
-export const marksSchema = z.object({
-	correct: z.number().int().default(1),
-	wrong: z.number().int().default(0),
-	skipped: z.number().int().default(0),
-	partial: z
-		.object({
-			correct_answers: z
-				.array(
-					z.object({
-						num_correct_selected: z.number().int().min(1),
-						marks: z.number().int().min(0)
-					})
-				)
-				.default([])
-		})
-		.optional()
-});
+import { marksSchema } from '$lib/schemas/marks';
 
 export enum MarksLevel {
 	QUESTION = 'question',
